salesforce-opengraph/requirements.txt
requests
PyJWT
yaml
